requirements.txt
httpx>=0.24.0
redis>=4.6.0
python-telegram-bot>=20.0a0
aiohttp>=3.8.0
python-dotenv>=1.0.0
